// types.ts
export interface BaseItem {
  id: string;
  name: string;
  description?: string;
}